Move SPIR-V util functions to a separate module.
